+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 * Fixed an issue where **validate** suggests, with no reason, running **format** on missing mandatory keys in yml file.
+* Skipped existence of TestPlaybook check on community and contribution integrations.
+* Fixed an issue where pre-commit didn't run on the demisto_sdk/commands folder.
 
 # 1.4.5
 * Enhanced the **postman-codegen** command to name all generated arguments with lower case.
@@ -12,6 +14,7 @@
 * The **format** command will now set `fromVersion: 6.2.0` for widgets with 'metrics' data type.
 * Updated the **find-dependencies** command to support generic modules, definitions, fields and types.
 * Fixed an issue where **openapi-codegen** tried to extract reference example outputs, leading to an exception.
+* Added an option to ignore secrets automatically when using the **init** command to create a pack.
 * Added a tool that gives the ability to temporarily suppress console output.
 * Added to **validate** command a validation that a content entity or a pack name does not contain the words "partner" and "community".