+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+
+# 9.9.9
 * Fixed an issue where **validate** suggests, with no reason, running **format** on missing mandatory keys in yml file.
 * Skipped existence of TestPlaybook check on community and contribution integrations.
 * Fixed an issue where pre-commit didn't run on the demisto_sdk/commands folder.